GreulerAdapter - translate between event coordinates and internal coords
The node coordinate properties are relative to the bounds of the svg
element. The event coordinate properties are relative to the screen
(clientX/Y are relative displayed content of the webpage).

Those two don't match up now that there's content on top of the graph.
So we need to translate the between the client coordinates and the
coordinates inside the svg in order to match nodes.