Component - use switch to handle mouse/touch modes
On mobile, both the mouse events and the touch events are triggered. We
want to handle the touch events, and if there was recently a touch,
handle touch end but not mouse up. The same goes for mouse events: if
there was a recent mouse event, handle mouse up but not touch end.

This follows a switch pattern: we go into mouse mode or touch mode, and
ignore the other mode.