ModeSwitch - add method to perform function if state is active
This facilitates a "tell, don't ask" interaction between clients and the
switch. Clients will tell the mode to do something instead of asking for
the current state and making a decision based on that. It also allows
the switch to encapsulate the state and clients to make decisions based
on that without asking for it. They will get it passed in to the
function they supply.

We need to accept all the modes at once rather than chaining in case one
of the actions transitions out of the current mode. We wouldn't want the
other actions to be performed in that case.