Graph - extract EditMode class for handling edit mode
`Graph` was having to maintain a bunch of disparate state to keep track
of edit mode, which is the first sign that this wanted to be its own
object. The next sign is that I want to be able to inject different edit
modes - a no-op edit mode, for instance, that never gets triggered and
so never allows connections to be made.