Mode State - add some logging
*If you tap into the state changes you tap into the heart of the system.*

This utility class has enabled some higher order thinking on my part, because
I can stop thinking about _doing_ the things of state manipulation and just
use the thing I made. Now I think in higher order terms, of _edit mode_
being composed of different modes, the modes for the individual labels, but
because there's a structured manner in which the two can communicate, it's
not a swamp of complexity.

squash! Mode State - add some logging