ModeSwitch - s/state/mode/
This has been a common conflation in my head. I initially started with
_Mode_, because I was thinking of different modes of using the device,
mouse and touch, and there wanted to be some way of switching between
the two. Modes felt natural to describe that state.

But now the graph is starting to git more complicated, so I'm going to
want to build standard use abstractions on top of modes. I got most
of the way through a concept called a _Bridge_, which contains two modes
and has state common to the two. That's a common patter amongst the
problenms I'm solving, from component at the event level to edit mode
at the state level.