Graph - split up subclasses into injected services
I want to be able to control different aspects of the graph differently;
e.g. set up a graph where I can add edges and change colors but not add
nodes, or set up a graph where all I can do is change colors. For that
flexibility we need to inject objects to perform those tasks, so we can
mix and match.

Most of these are basically typed functions. Making them objects
provides the advantage of being able to generate code for them.