GraphComponent - remove direct dependence on adapter
The `GraphComponent` was interfacing with `Graph` and also directly with
`GreulerAdapter` to call `getClickTarget`. That's not desirable because
it creates a more complex triangle dependency diagram. The reason I did
it initially was that the method was called `getClickTarget` and took in
an event object directly, whereas I wanted `Graph` to be uncoupled from
concerns about the DOM. `Graph` was supposed to be about the logical
data structure that was the graph and not know about the details of
being embedded in the DOM. However, that created a dependency graph that
reached across abstraction boundaries and was more complicated than
necessary. Instead, we should reframe `getClickTarget` as
`getNearestElement` so that it's only about graph elements.